refactor: VO 클래스를 request 패키지와 response 패키지로 분리하기
