+++ src/test/java/lotto/utility/vo/response/ProfitResponseTest.java
@@ -1,8 +1,9 @@
-package lotto.utility.vo;
+package lotto.utility.vo.response;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.stream.Stream;
+import lotto.utility.vo.response.ProfitResponse;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
